Extracted email-related functionality out of base comments service file
refs https://github.com/TryGhost/Team/issues/1664
