+++ core/server/services/comments/email-templates/new-comment.hbs
