+++ core/server/services/comments/email-templates/new-comment.txt.js
